Changed the API up in preparation for SSL work
Separated context intialization from binding listener socket.
Also added libwebsock_event_container to determine if epoll event came from listener or client.
libwebsock_listener_state->flags will now tell us if the client connected to an SSL enabled port
and handle that appropriately.  The actual handling of this is upcoming.
This allows one to bind the server to many specific addresses/ports if need be.
Also lays the groundwork for libwebsock_bind_ssl which will be similar to libwebsock_bind
but will flag the listener as SSL enabled and accepted connections to an SSL listener
will be handled with new SSL functions.  Those are on the way.